+++ src/Git_Updater/Base.php
@@ -345,8 +345,9 @@ class Base {
 			$ran[] = false !== $repo_api->get_remote_branches()  ? 'branches' : null;
 			$ran[] = false !== $repo_api->get_repo_meta()        ? 'meta'     : null;
 			$repo_api->set_repo_cache( 'ran', array_filter( $ran ) );
-			$repo_api->set_repo_cache_timeout( $repo->slug );
 			// phpcs:enable
+			$repo->download_link = $repo_api->construct_download_link();
+			$repo_api->set_repo_cache_timeout( $repo->slug );
 			$language_pack = new Language_Pack( $repo, new Language_Pack_API( $repo ) );
 			$language_pack->run();
 		}